% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_poly.R
\name{BB_poly}
\alias{BB_poly}
\title{BB_poly}
\usage{
BB_poly(model, ..., data, criterion, max_mu = 4, max_sigma = 4)
}
\arguments{
\item{model}{A formula that defines the model for continuous norming.}

\item{data}{Defines the data set.}

\item{criterion}{Either "BIC", "AIC" or "GAIC(3)".}

\item{max.mu}{An integer. Defines the maximum power for polynomials.}

\item{max.sigma}{An integer. Defines the maximum power for polynomials.}
}
\value{
The gamlss model selected by the free order procedure.
}
\description{
This function models norming with a beta-binomial distribution.
}
\details{
This function uses polynomials.
}
\examples{
# Apply the free order procedure for a BCPE and the BIC.
BB_poly(model = g ~ Normgruppe,
        data = Data_norm,
        criterion = "BIC")

}
\author{
Julian Urban
}
