% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_pb.R
\name{gamlss_pb}
\alias{gamlss_pb}
\title{gamlss_pb}
\usage{
gamlss_pb(model, ..., data, family, criterion)
}
\arguments{
\item{model}{A formula that defines the model for continuous norming.}

\item{data}{Defines the data set.}

\item{family}{A character that defines the specific PDF for modelling.}

\item{criterion}{Either "BIC", "AIC" or "GAIC(3)".}
}
\value{
The gamlss model selected by the free order procedure.
}
\description{
This function is in inner function and conducts the free order procedure
with penalized splines.
}
\details{
This function is used for penalized splines
}
\examples{
# Apply the free order procedure for a BCPE and the BIC.
gamlss_pb(model = g ~ Normgruppe,
          data = Data_norm,
          criterion = "BIC",
          family = "NO")

}
\author{
Julian Urban
}
