% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_poly.R
\name{gamlss_poly}
\alias{gamlss_poly}
\title{gamlss_poly}
\usage{
gamlss_poly(
  model,
  ...,
  data,
  family,
  criterion,
  max.mu = max.mu,
  max.sigma = max.sigma,
  max.nu = max.nu,
  max.tau = max.tau
)
}
\arguments{
\item{model}{A formula that defines the model for continuous norming.}

\item{data}{Defines the data set.}

\item{family}{A character that defines the specific PDF for modelling.}

\item{criterion}{Either "BIC", "AIC" or "GAIC(3)".}

\item{max.mu}{An integer. Defines the maximum power for polynomials.}

\item{max.sigma}{An integer. Defines the maximum power for polynomials.}

\item{max.nu}{An integer. Defines the maximum power for polynomials.}

\item{max.tau}{An integer. Defines the maximum power for polynomials.}
}
\value{
The gamlss model selected by the free order procedure.
}
\description{
This function is in inner function and conducts the free order procedure
with polynomials.
}
\details{
This function is used for polynomials.
}
\examples{
# Apply the free order procedure for a BCPE and the BIC.
gamlss_poly(model = g ~ Normgruppe,
            data = Data_norm,
            criterion = "BIC",
            family = "BCPW")

}
\author{
Julian Urban
}
