% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_wrapper.R
\name{gamlss_wrapper}
\alias{gamlss_wrapper}
\title{gamlss_wrapper}
\usage{
gamlss_wrapper(
  model,
  ...,
  data,
  family,
  smooth,
  criterion,
  max.mu = 20,
  max.sigma = 20,
  max.nu = 20,
  max.tau = 20
)
}
\arguments{
\item{model}{A formula that defines the model for continuous norming.}

\item{data}{Defines the data set.}

\item{family}{A character that defines the specific PDF for modelling.}

\item{smooth}{A character defining the smooth function.}

\item{criterion}{Either "BIC", "AIC" or "GAIC(3)".}

\item{max.mu}{An integer. Defines the maximum power for polynomials.}

\item{max.sigma}{An integer. Defines the maximum power for polynomials.}

\item{max.nu}{An integer. Defines the maximum power for polynomials.}

\item{max.tau}{An integer. Defines the maximum power for polynomials.}
}
\value{
The gamlss model selected by the free order procedure.
}
\description{
This function conducts the free order procedure for any PDF using either
power polynomials or penalized splines
}
\details{
This function is used to select the model.
}
\examples{
# Apply the free order procedure for a BCPE and the BIC.
gamlss_wrapper(model = g ~ e_age_month,
               data = Data_norm,
               criterion = "BIC",
               family = "NO",
               smooth = "pb")

}
\author{
Julian Urban
}
