% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gamlss_norms.R
\name{para_norm_extraction}
\alias{para_norm_extraction}
\title{para_norm_extraction}
\usage{
para_norm_extraction(model, scale = "percentiles")
}
\arguments{
\item{model}{GAMLSS model for which norm scores should be extracted}

\item{scale}{Defines the scale of the norm scores}
}
\value{
The norm scores
}
\description{
This function extracts the norm scores of a gamlss norming model.
}
\details{
This function provide norm scores.
}
\author{
Julian Urban
}
