% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{str_eval}
\alias{str_eval}
\title{str_eval}
\usage{
str_eval(x)
}
\arguments{
\item{x}{Model to be evaluates}
}
\value{
model
}
\description{
This function is an inner function to evaluate the models.
}
\details{
This function is used within the free order procedure
}
\author{
Julian Urban
}
