% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_poly.R
\name{three_parameters_poly}
\alias{three_parameters_poly}
\title{three_parameters_poly}
\usage{
three_parameters_poly(
  model,
  ...,
  data,
  family,
  criterion,
  max.mu = max.mu,
  max.sigma = max.sigma,
  max.nu = max.nu
)
}
\arguments{
\item{model}{A formula that defines the model for continuous norming.}

\item{data}{Defines the data set.}

\item{family}{A character that defines the specific PDF for modelling.}

\item{criterion}{Either "BIC", "AIC" or "GAIC(3)".}

\item{max.mu}{An integer. Defines the maximum power for polynomials.}

\item{max.sigma}{An integer. Defines the maximum power for polynomials.}

\item{max.nu}{An integer. Defines the maximum power for polynomials.}
}
\value{
The gamlss model selected by the free order procedure.
}
\description{
This function is in inner function and conducts the free order procedure
with polynomials for PDFs with three parameters.
}
\details{
This function is used for polynomials.
}
\examples{
# Apply the free order procedure for a BCPE and the BIC.
three_parameters_ploy(model = g ~ Normgruppe,
                     data = Data_norm,
                     criterion = "BIC",
                     family = "ST1")

}
\author{
Julian Urban
}
